Don't add tests in `__init__`
